+++ example-context/example-service/example-command/example-command-api/src/test/java/uk/gov/justice/services/example/cakeshop/command/api/OrderCakeCommandApiTest.java
@@ -1,7 +1,10 @@
 package uk.gov.justice.services.example.cakeshop.command.api;
 
 
-import static uk.gov.justice.services.test.utils.core.helper.ServiceComponents.verifyPassThroughCommandHandlerMethod;
+import static org.junit.Assert.assertThat;
+import static uk.gov.justice.services.core.annotation.Component.COMMAND_API;
+import static uk.gov.justice.services.test.utils.core.matchers.HandlerClassMatcher.isHandlerClass;
+import static uk.gov.justice.services.test.utils.core.matchers.HandlerMethodMatcher.method;
 
 import org.junit.Test;
 
@@ -9,6 +12,9 @@ public class OrderCakeCommandApiTest {
 
     @Test
     public void shouldHandleOrderCakeCommand() throws Exception {
-        verifyPassThroughCommandHandlerMethod(OrderCakeCommandApi.class);
+        assertThat(OrderCakeCommandApi.class, isHandlerClass(COMMAND_API)
+                .with(method("orderCake")
+                        .thatHandles("example.order-cake")
+                        .withSenderPassThrough()));
     }
 }