+++ example-context/example-service/example-command/example-command-handler/src/test/java/uk/gov/justice/services/example/cakeshop/command/handler/MakeCakeCommandHandlerTest.java
@@ -1,7 +1,11 @@
 package uk.gov.justice.services.example.cakeshop.command.handler;
 
+import static org.hamcrest.MatcherAssert.assertThat;
+import static uk.gov.justice.services.core.annotation.Component.COMMAND_HANDLER;
 import static uk.gov.justice.services.messaging.DefaultJsonEnvelope.envelope;
 import static uk.gov.justice.services.messaging.JsonObjectMetadata.metadataWithRandomUUID;
+import static uk.gov.justice.services.test.utils.core.matchers.HandlerMatcher.isHandler;
+import static uk.gov.justice.services.test.utils.core.matchers.HandlerMethodMatcher.method;
 
 import uk.gov.justice.services.eventsourcing.source.core.EventSource;
 import uk.gov.justice.services.eventsourcing.source.core.EventStream;
@@ -34,8 +38,14 @@ public class MakeCakeCommandHandlerTest {
     private MakeCakeCommandHandler makeCakeCommandHandler;
 
     @Test
+    public void shouldHaveCorrectHandlesAnnotation() throws Exception {
+        assertThat(makeCakeCommandHandler, isHandler(COMMAND_HANDLER)
+                .with(method("makeCake").thatHandles("example.make-cake")));
+    }
+
+    @Test
     public void shouldHandleMakeCakeCommand() throws Exception {
-        makeCakeCommandHandler.handle(
+        makeCakeCommandHandler.makeCake(
                 envelope()
                         .with(metadataWithRandomUUID(EVENT_NAME))
                         .withPayloadOf(CAKE_ID.toString(), "cakeId")