+++ example-context/example-service/example-query/example-query-api/src/test/java/uk/gov/justice/services/example/cakeshop/query/api/CakeOrdersQueryApiTest.java
@@ -1,7 +1,10 @@
 package uk.gov.justice.services.example.cakeshop.query.api;
 
 
-import static uk.gov.justice.services.test.utils.core.helper.ServiceComponents.verifyPassThroughQueryHandlerMethod;
+import static org.junit.Assert.assertThat;
+import static uk.gov.justice.services.core.annotation.Component.QUERY_API;
+import static uk.gov.justice.services.test.utils.core.matchers.HandlerClassMatcher.isHandlerClass;
+import static uk.gov.justice.services.test.utils.core.matchers.HandlerMethodMatcher.method;
 
 import org.junit.Test;
 
@@ -9,6 +12,9 @@ public class CakeOrdersQueryApiTest {
 
     @Test
     public void shouldHandleCakeOrderQuery() throws Exception {
-        verifyPassThroughQueryHandlerMethod(CakeOrdersQueryApi.class);
+        assertThat(CakeOrdersQueryApi.class, isHandlerClass(QUERY_API)
+                .with(method("getOrder")
+                        .thatHandles("example.get-order")
+                        .withRequesterPassThrough()));
     }
 }