+++ example-context/example-service/example-query/example-query-api/src/test/java/uk/gov/justice/services/example/cakeshop/query/api/RecipesQueryApiTest.java
@@ -1,14 +1,26 @@
 package uk.gov.justice.services.example.cakeshop.query.api;
 
 
-import static uk.gov.justice.services.test.utils.core.helper.ServiceComponents.verifyPassThroughQueryHandlerMethod;
+import static org.hamcrest.Matchers.allOf;
+import static org.junit.Assert.assertThat;
+import static uk.gov.justice.services.core.annotation.Component.QUERY_API;
+import static uk.gov.justice.services.test.utils.core.matchers.HandlerClassMatcher.isHandlerClass;
+import static uk.gov.justice.services.test.utils.core.matchers.HandlerMethodMatcher.method;
 
 import org.junit.Test;
 
 public class RecipesQueryApiTest {
 
     @Test
-    public void shouldHandleRecipesQuery() throws Exception {
-        verifyPassThroughQueryHandlerMethod(RecipesQueryApi.class);
+    public void shouldBeQueryApiThatHasRequesterPassThroughMethods() throws Exception {
+        assertThat(RecipesQueryApi.class, isHandlerClass(QUERY_API)
+                .with(allOf(
+                        method("searchRecipes")
+                                .thatHandles("example.search-recipes")
+                                .withRequesterPassThrough(),
+                        method("getRecipe")
+                                .thatHandles("example.get-recipe")
+                                .withRequesterPassThrough()
+                )));
     }
 }