+++ example-context/example-service/example-query/example-query-controller/src/test/java/uk/gov/justice/services/example/cakeshop/query/controller/CakeOrdersQueryControllerTest.java
@@ -1,6 +1,9 @@
 package uk.gov.justice.services.example.cakeshop.query.controller;
 
-import static uk.gov.justice.services.test.utils.core.helper.ServiceComponents.verifyPassThroughQueryHandlerMethod;
+import static org.junit.Assert.assertThat;
+import static uk.gov.justice.services.core.annotation.Component.QUERY_CONTROLLER;
+import static uk.gov.justice.services.test.utils.core.matchers.HandlerClassMatcher.isHandlerClass;
+import static uk.gov.justice.services.test.utils.core.matchers.HandlerMethodMatcher.method;
 
 import org.junit.Test;
 
@@ -8,6 +11,9 @@ public class CakeOrdersQueryControllerTest {
 
     @Test
     public void shouldHandleCakeOrdersQuery() throws Exception {
-        verifyPassThroughQueryHandlerMethod(CakeOrdersQueryController.class);
+        assertThat(CakeOrdersQueryController.class, isHandlerClass(QUERY_CONTROLLER)
+                .with(method("getOrder")
+                        .thatHandles("example.get-order")
+                        .withRequesterPassThrough()));
     }
 }