+++ example-context/example-service/example-query/example-query-view/src/test/java/uk/gov/justice/services/example/cakeshop/query/view/CakeOrdersQueryViewTest.java
@@ -4,8 +4,11 @@ package uk.gov.justice.services.example.cakeshop.query.view;
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.when;
+import static uk.gov.justice.services.core.annotation.Component.QUERY_VIEW;
 import static uk.gov.justice.services.messaging.DefaultJsonEnvelope.envelope;
 import static uk.gov.justice.services.messaging.JsonObjectMetadata.metadataWithDefaults;
+import static uk.gov.justice.services.test.utils.core.matchers.HandlerMatcher.isHandler;
+import static uk.gov.justice.services.test.utils.core.matchers.HandlerMethodMatcher.method;
 
 import uk.gov.justice.services.common.converter.ObjectToJsonValueConverter;
 import uk.gov.justice.services.common.converter.jackson.ObjectMapperProducer;
@@ -35,6 +38,12 @@ public class CakeOrdersQueryViewTest {
     public void setup() {
         final Enveloper enveloper = new Enveloper(new ObjectToJsonValueConverter(new ObjectMapperProducer().objectMapper()));
         queryView = new CakeOrdersQueryView(service, enveloper);
+    }
+
+    @Test
+    public void shouldHaveCorrectHandlerMethod() throws Exception {
+        assertThat(queryView, isHandler(QUERY_VIEW)
+                .with(method("findOrder").thatHandles("example.get-order")));
     }
 
     @Test